+++ DomainManagement/functions/gplinks/Unregister-DMGPLink.ps1
@@ -18,21 +18,36 @@
 
 		Clears all configured Group policy links.
 	#>
-	[CmdletBinding()]
+	[CmdletBinding(DefaultParameterSetName = 'Path')]
 	param (
 		[parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
 		[string]
 		$PolicyName,
 
-		[parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
+		[parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'Path')]
 		[Alias('OU')]
 		[string]
-		$OrganizationalUnit
+		$OrganizationalUnit,
+
+		[parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'Filter')]
+		[string]
+		$OUFilter
 	)
 	
 	process
 	{
-		$script:groupPolicyLinks[$OrganizationalUnit].Remove($PolicyName)
-		if ($script:groupPolicyLinks[$OrganizationalUnit].Keys.Count -lt 1) { $script:groupPolicyLinks.Remove($OrganizationalUnit) }
+		switch ($PSCmdlet.ParameterSetName) {
+			'Path'
+			{
+				$script:groupPolicyLinks[$OrganizationalUnit].Remove($PolicyName)
+				if ($script:groupPolicyLinks[$OrganizationalUnit].Keys.Count -lt 1) { $script:groupPolicyLinks.Remove($OrganizationalUnit) }
+			}
+			'Filter'
+			{
+				$script:groupPolicyLinksDynamic[$OUFilter].Remove($PolicyName)
+				if ($script:groupPolicyLinksDynamic[$OUFilter].Keys.Count -lt 1) { $script:groupPolicyLinksDynamic.Remove($OUFilter) }
+			}
+		}
+		
 	}
 }